app/requirements.txt
annotated-types==0.7.0
APScheduler==3.10.4
astroid==3.2.4
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
Flask==3.0.3
google-auth==2.29.0
gunicorn==22.0.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
kubernetes==29.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==4.2.2
prometheus_client==0.20.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.10.6
pydantic-core==2.27.2
pylint==3.2.6
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tomlkit==0.13.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.6